requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==3.0.4
idna==2.8
Pillow==8.3.1
requests==2.22.0
soupsieve==2.2.1
tqdm==4.61.2
urllib3==1.25.11
